+++ src/client.rs
@@ -8,8 +8,8 @@ pub use config::*;
 pub(crate) use connection::*;
 
 use crate::{
-    result::{ExecuteResult, QueryResult},
-    tds::{codec, stream::TokenStream},
+    result::ExecuteResult,
+    tds::{codec, stream::{TokenStream, QueryStream}},
     SqlReadBytes, ToSql,
 };
 use codec::{BatchRequest, ColumnData, PacketHeader, RpcParam, RpcProcId, TokenRpcRequest};
@@ -119,7 +119,7 @@ impl<S: AsyncRead + AsyncWrite + Unpin + Send> Client<S> {
     /// Useful for `SELECT` statements. The `query` can define the parameter
     /// placement by annotating them with `@PN`, where N is the index of the
     /// parameter, starting from `1`. If executing multiple queries at a time,
-    /// delimit them with `;` and refer to [`QueryResult`] on proper stream
+    /// delimit them with `;` and refer to [`QueryStream`] on proper stream
     /// handling.
     ///
     /// For mapping of Rust types when writing, see the documentation for
@@ -151,14 +151,14 @@ impl<S: AsyncRead + AsyncWrite + Unpin + Send> Client<S> {
     /// # }
     /// ```
     ///
-    /// [`QueryResult`]: struct.QueryResult.html
+    /// [`QueryStream`]: struct.QueryStream.html
     /// [`ToSql`]: trait.ToSql.html
     /// [`FromSql`]: trait.FromSql.html
     pub async fn query<'a, 'b>(
         &'a mut self,
         query: impl Into<Cow<'b, str>>,
         params: &'b [&'b dyn ToSql],
-    ) -> crate::Result<QueryResult<'a>>
+    ) -> crate::Result<QueryStream<'a>>
     where
         'a: 'b,
     {
@@ -169,7 +169,7 @@ impl<S: AsyncRead + AsyncWrite + Unpin + Send> Client<S> {
             .await?;
 
         let ts = TokenStream::new(&mut self.connection);
-        let mut result = QueryResult::new(ts.try_unfold());
+        let mut result = QueryStream::new(ts.try_unfold());
         result.forward_to_metadata().await?;
 
         Ok(result)
@@ -208,7 +208,7 @@ impl<S: AsyncRead + AsyncWrite + Unpin + Send> Client<S> {
     pub async fn simple_query<'a, 'b>(
         &'a mut self,
         query: impl Into<Cow<'b, str>>,
-    ) -> crate::Result<QueryResult<'a>>
+    ) -> crate::Result<QueryStream<'a>>
     where
         'a: 'b,
     {
@@ -221,7 +221,7 @@ impl<S: AsyncRead + AsyncWrite + Unpin + Send> Client<S> {
 
         let ts = TokenStream::new(&mut self.connection);
 
-        let mut result = QueryResult::new(ts.try_unfold());
+        let mut result = QueryStream::new(ts.try_unfold());
         result.forward_to_metadata().await?;
 
         Ok(result)
