+++ src/lib.rs
@@ -196,7 +196,7 @@ pub use from_sql::{FromSql, FromSqlOwned};
 pub use result::*;
 pub use row::{Column, ColumnType, Row};
 pub use sql_browser::SqlBrowser;
-pub use tds::{codec::ColumnData, numeric, time, xml, EncryptionLevel};
+pub use tds::{codec::ColumnData, numeric, stream::QueryStream, time, xml, EncryptionLevel};
 pub use to_sql::{IntoSql, ToSql};
 pub use uuid::Uuid;
