+++ src/result.rs
@@ -1,221 +1,10 @@
 pub use crate::tds::stream::{QueryItem, ResultMetadata};
 use crate::{
     client::Connection,
-    tds::stream::{QueryStream, ReceivedToken, TokenStream},
-    Column, Row,
+    tds::stream::{ReceivedToken, TokenStream},
 };
-use futures::{stream::BoxStream, AsyncRead, AsyncWrite, Stream, TryStreamExt};
-use std::{fmt::Debug, pin::Pin, task};
-use task::Poll;
-
-/// A set of `Streams` of [`Rows`] resulting from a `SELECT` query. The
-/// `QueryResult` needs to be polled empty before sending another query to the
-/// [`Client`], failing to do so causes a flush before the next query, slowing it
-/// down in an undeterministic way.
-///
-/// A stream consists of [`QueryItem`] values, which can be either result
-/// metadata or a row. Every stream starts with metadata, describing the
-/// structure of the incoming rows, e.g. the columns in the order they are
-/// presented in every row.
-///
-/// If after consuming rows from the stream, another metadata result arrives, it
-/// means the stream has multiple results from different queries. This new
-/// metadata item will describe the next rows from here forwards.
-///
-/// # Example
-///
-/// ```
-/// # use tiberius::{Config, QueryItem};
-/// # use tokio_util::compat::TokioAsyncWriteCompatExt;
-/// # use std::env;
-/// # use futures::TryStreamExt;
-/// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// # let c_str = env::var("TIBERIUS_TEST_CONNECTION_STRING").unwrap_or(
-/// #     "server=tcp:localhost,1433;integratedSecurity=true;TrustServerCertificate=true".to_owned(),
-/// # );
-/// # let config = Config::from_ado_string(&c_str)?;
-/// # let tcp = tokio::net::TcpStream::connect(config.get_addr()).await?;
-/// # tcp.set_nodelay(true)?;
-/// # let mut client = tiberius::Client::connect(config, tcp.compat_write()).await?;
-/// let mut stream = client
-///     .query(
-///         "SELECT @P1 AS first; SELECT @P2 AS second",
-///         &[&1i32, &2i32],
-///     )
-///     .await?;
-///
-/// // The stream consists of four items, in the following order:
-/// // - Metadata from `SELECT 1`
-/// // - The only resulting row from `SELECT 1`
-/// // - Metadata from `SELECT 2`
-/// // - The only resulting row from `SELECT 2`
-/// while let Some(item) = stream.try_next().await? {
-///     match item {
-///         // our first item is the column data always
-///         QueryItem::Metadata(meta) if meta.result_index() == 0 => {
-///             // the first result column info can be handled here
-///         }
-///         // ... and from there on from 0..N rows
-///         QueryItem::Row(row) if row.result_index() == 0 => {
-///             assert_eq!(Some(1), row.get(0));
-///         }
-///         // the second result set returns first another metadata item
-///         QueryItem::Metadata(meta) => {
-///             // .. handling
-///         }
-///         // ...and, again, we get rows from the second resultset
-///         QueryItem::Row(row) => {
-///             assert_eq!(Some(2), row.get(0));
-///         }
-///     }
-/// }
-/// # Ok(())
-/// # }
-/// ```
-///
-/// [`Client`]: struct.Client.html
-/// [`Rows`]: struct.Row.html
-/// [`next_resultset`]: #method.next_resultset
-#[derive(Debug)]
-pub struct QueryResult<'a> {
-    stream: QueryStream<'a>,
-}
-
-impl<'a> QueryResult<'a> {
-    pub(crate) fn new(token_stream: BoxStream<'a, crate::Result<ReceivedToken>>) -> Self {
-        let stream = QueryStream::new(token_stream);
-
-        Self { stream }
-    }
-
-    pub(crate) async fn forward_to_metadata(&mut self) -> crate::Result<()> {
-        self.stream.forward_to_metadata().await
-    }
-
-    /// The list of columns either for the current result set, or for the next
-    /// one. If the stream is just created, or if the next item in the stream
-    /// contains metadata, the metadata will be taken from the stream. Otherwise
-    /// the columns will be returned from the cache and reflect on the current
-    /// result set.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use tiberius::Config;
-    /// # use tokio_util::compat::TokioAsyncWriteCompatExt;
-    /// # use std::env;
-    /// # use futures::TryStreamExt;
-    /// # #[tokio::main]
-    /// # async fn main() -> anyhow::Result<()> {
-    /// # let c_str = env::var("TIBERIUS_TEST_CONNECTION_STRING").unwrap_or(
-    /// #     "server=tcp:localhost,1433;integratedSecurity=true;TrustServerCertificate=true".to_owned(),
-    /// # );
-    /// # let config = Config::from_ado_string(&c_str)?;
-    /// # let tcp = tokio::net::TcpStream::connect(config.get_addr()).await?;
-    /// # tcp.set_nodelay(true)?;
-    /// # let mut client = tiberius::Client::connect(config, tcp.compat_write()).await?;
-    /// let mut stream = client
-    ///     .query(
-    ///         "SELECT @P1 AS first; SELECT @P2 AS second",
-    ///         &[&1i32, &2i32],
-    ///     )
-    ///     .await?;
-    ///
-    /// // Nothing is fetched, the first result set starts.
-    /// let cols = stream.columns().await?.unwrap();
-    /// assert_eq!("first", cols[0].name());
-    ///
-    /// // Move over the metadata.
-    /// stream.try_next().await?;
-    ///
-    /// // We're in the first row, seeing the metadata for that set.
-    /// let cols = stream.columns().await?.unwrap();
-    /// assert_eq!("first", cols[0].name());
-    ///
-    /// // Move over the only row in the first set.
-    /// stream.try_next().await?;
-    ///
-    /// // End of the first set, getting the metadata by peaking the next item.
-    /// let cols = stream.columns().await?.unwrap();
-    /// assert_eq!("second", cols[0].name());
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub async fn columns(&mut self) -> crate::Result<Option<&[Column]>> {
-        self.stream.metadata().await
-    }
-
-    /// Collects results from all queries in the stream into memory in the order
-    /// of querying.
-    pub async fn into_results(mut self) -> crate::Result<Vec<Vec<Row>>> {
-        let mut results: Vec<Vec<Row>> = Vec::new();
-        let mut result: Option<Vec<Row>> = None;
-
-        while let Some(item) = self.stream.try_next().await? {
-            match (item, &mut result) {
-                (QueryItem::Row(row), None) => {
-                    result.insert({
-                        let mut result = Vec::new();
-                        result.push(row);
-                        result
-                    });
-                }
-                (QueryItem::Row(row), Some(ref mut result)) => result.push(row),
-                (QueryItem::Metadata(_), None) => {
-                    result.insert(Vec::new());
-                }
-                (QueryItem::Metadata(_), ref mut previous_result) => {
-                    results.push(previous_result.take().unwrap());
-                    result = None;
-                }
-            }
-        }
-
-        if let Some(result) = result {
-            results.push(result);
-        }
-
-        Ok(results)
-    }
-
-    /// Collects the output of the first query, dropping any further
-    /// results.
-    pub async fn into_first_result(self) -> crate::Result<Vec<Row>> {
-        let mut results = self.into_results().await?.into_iter();
-        let rows = results.next().unwrap_or_else(Vec::new);
-
-        Ok(rows)
-    }
-
-    /// Collects the first row from the output of the first query, dropping any
-    /// further rows.
-    pub async fn into_row(self) -> crate::Result<Option<Row>> {
-        let mut results = self.into_first_result().await?.into_iter();
-
-        Ok(results.next())
-    }
-
-    /// Convert the stream into a stream of rows, skipping metadata items.
-    pub fn into_row_stream(self) -> BoxStream<'a, crate::Result<Row>> {
-        let s = self.stream.try_filter_map(|item| async {
-            match item {
-                QueryItem::Row(row) => Ok(Some(row)),
-                QueryItem::Metadata(_) => Ok(None),
-            }
-        });
-
-        Box::pin(s)
-    }
-}
-
-impl<'a> Stream for QueryResult<'a> {
-    type Item = crate::Result<QueryItem>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Option<Self::Item>> {
-        Pin::new(&mut self.get_mut().stream).poll_next(cx)
-    }
-}
+use futures::{AsyncRead, AsyncWrite, TryStreamExt};
+use std::fmt::Debug;
 
 /// A result from a query execution, listing the number of affected rows.
 ///
